requirements.txt
stomp.py
dotenv